Inline ssh-store.hh into ssh-store.cc
